ocenianie kolosow

+++ api/src/main/java/com/example/api/activity/result/dto/request/AddColloquiumPointsForm.java
@@ -15,6 +15,6 @@ public class AddColloquiumPointsForm {
     @Schema(required = true) private Long colloquiumId;
     @Schema(required = false) private String description;
     @Schema(required = true) private Long annihilatedQuestions;
-    @Schema(required = true) private Long annihilatedPoints;
+    @Schema(required = true) private Double annihilatedPoints;
     @Schema(required = true) private Long dateInMillis;
 }
